+++ index.html
@@ -247,8 +247,8 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-  <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
+
+  <script src="js/main.js" type="module"></script>
 
 </body>
